the cancel order now works, and i fixed the total price in the order …

+++ server/src/main/java/il/cshaifasweng/OCSFMediatorExample/server/mapping/OrderMapper.java
@@ -3,34 +3,32 @@ package il.cshaifasweng.OCSFMediatorExample.server.mapping;
 import il.cshaifasweng.OCSFMediatorExample.entities.domain.Order;
 import il.cshaifasweng.OCSFMediatorExample.entities.messages.CheckOut.OrderDTO;
 
+import java.math.BigDecimal;
 import java.util.stream.Collectors;
 
 public class OrderMapper {
 
     public static Order fromDTO(OrderDTO dto) {
         if (dto == null) return null;
-
         Order order = new Order();
         order.setId(dto.getId());
         order.setCustomerId(dto.getCustomerId());
         order.setStoreId(dto.getStoreId());
         order.setCreatedAt(dto.getCreatedAt());
 
-// Status: map by name if entity uses a different enum type
+        // Map status safely
         if (dto.getStatus() != null) {
             try {
                 order.setStatus(Order.Status.valueOf(dto.getStatus().name()));
             } catch (IllegalArgumentException ex) {
-                // pick a sane default
                 order.setStatus(Order.Status.INITIATED);
             }
         }
 
-
-        // Totals from client (will be recomputed in handler anyway)
+        // Monetary totals — convert safely to BigDecimal
         order.setSubtotal(dto.getSubtotal());
         order.setDiscountTotal(dto.getDiscountTotal());
-        order.setTotal(dto.getTotal());
+        order.setTotal(BigDecimal.valueOf(dto.getTotal()));
 
         // Nested objects
         order.setDelivery(DeliveryInfoMapper.fromDTO(dto.getDelivery()));
